fix(mobile): show composer menus when starting a task
